Revert "fix Z3_Backend.hpp(320): warning C4244: 'initializing': conversion from 'const unsigned __int64' to 'unsigned int'"
This reverts commit a330976c3630dfa7ef4e1f952ccc57ea14d48e76.

+++ src/metaSMT/backend/Z3_Backend.hpp
@@ -314,7 +314,7 @@ namespace metaSMT {
       }
 
       result_type operator() (bvtags::bvuint_tag const &, boost::any const &arg) {
-        typedef boost::tuple<uint64_t, unsigned> P;
+        typedef boost::tuple<uint64_t, uint64_t> P;
         P const p = boost::any_cast<P>(arg);
         uint64_t const value = boost::get<0>(p);
         unsigned const width = boost::get<1>(p);
